[BUILD] Enforce package naming conventions PMD
