+++ eclipse/src/saros/ui/command_handlers/AddContactHandler.java
@@ -1,17 +1,15 @@
-package saros.ui.commandHandlers;
+package saros.ui.command_handlers;
 
 import org.eclipse.core.commands.AbstractHandler;
 import org.eclipse.core.commands.ExecutionEvent;
 import org.eclipse.core.commands.ExecutionException;
 import saros.ui.util.WizardUtils;
 
-public class CreateXMPPAccountHandler extends AbstractHandler {
+public class AddContactHandler extends AbstractHandler {
 
   @Override
   public Object execute(ExecutionEvent event) throws ExecutionException {
-
-    WizardUtils.openCreateXMPPAccountWizard(true);
-
+    WizardUtils.openAddContactWizard();
     return null;
   }
 }
