+++ eclipse/src/saros/ui/command_handlers/AddXMPPAccountHandler.java
@@ -1,15 +1,15 @@
-package saros.ui.commandHandlers;
+package saros.ui.command_handlers;
 
 import org.eclipse.core.commands.AbstractHandler;
 import org.eclipse.core.commands.ExecutionEvent;
 import org.eclipse.core.commands.ExecutionException;
 import saros.ui.util.WizardUtils;
 
-public class AddContactHandler extends AbstractHandler {
+public class AddXMPPAccountHandler extends AbstractHandler {
 
   @Override
   public Object execute(ExecutionEvent event) throws ExecutionException {
-    WizardUtils.openAddContactWizard();
+    WizardUtils.openAddXMPPAccountWizard();
     return null;
   }
 }
