+++ eclipse/src/saros/ui/command_handlers/CreateXMPPAccountHandler.java
@@ -1,17 +1,16 @@
-package saros.ui.commandHandlers;
+package saros.ui.command_handlers;
 
 import org.eclipse.core.commands.AbstractHandler;
 import org.eclipse.core.commands.ExecutionEvent;
 import org.eclipse.core.commands.ExecutionException;
-import saros.ui.util.SWTUtils;
+import saros.ui.util.WizardUtils;
 
-public class GettingStartedHandler extends AbstractHandler {
+public class CreateXMPPAccountHandler extends AbstractHandler {
 
   @Override
   public Object execute(ExecutionEvent event) throws ExecutionException {
 
-    SWTUtils.openInternalBrowser(
-        "https://www.saros-project.org/documentation/getting-started.html", "Welcome to Saros");
+    WizardUtils.openCreateXMPPAccountWizard(true);
 
     return null;
   }
