+++ eclipse/src/saros/ui/command_handlers/GettingStartedHandler.java
@@ -1,15 +1,18 @@
-package saros.ui.commandHandlers;
+package saros.ui.command_handlers;
 
 import org.eclipse.core.commands.AbstractHandler;
 import org.eclipse.core.commands.ExecutionEvent;
 import org.eclipse.core.commands.ExecutionException;
-import saros.ui.util.WizardUtils;
+import saros.ui.util.SWTUtils;
 
-public class StartSarosConfigurationHandler extends AbstractHandler {
+public class GettingStartedHandler extends AbstractHandler {
 
   @Override
   public Object execute(ExecutionEvent event) throws ExecutionException {
-    WizardUtils.openSarosConfigurationWizard();
+
+    SWTUtils.openInternalBrowser(
+        "https://www.saros-project.org/documentation/getting-started.html", "Welcome to Saros");
+
     return null;
   }
 }
