+++ eclipse/src/saros/ui/command_handlers/StartSarosConfigurationHandler.java
@@ -1,15 +1,15 @@
-package saros.ui.commandHandlers;
+package saros.ui.command_handlers;
 
 import org.eclipse.core.commands.AbstractHandler;
 import org.eclipse.core.commands.ExecutionEvent;
 import org.eclipse.core.commands.ExecutionException;
-import saros.ui.util.CollaborationUtils;
+import saros.ui.util.WizardUtils;
 
-public class SessionLeaveHandler extends AbstractHandler {
+public class StartSarosConfigurationHandler extends AbstractHandler {
 
   @Override
   public Object execute(ExecutionEvent event) throws ExecutionException {
-    CollaborationUtils.leaveSession();
+    WizardUtils.openSarosConfigurationWizard();
     return null;
   }
 }
