+++ eclipse/src/saros/ui/preference_pages/FeedbackPreferencePage.java
@@ -1,5 +1,5 @@
 /** */
-package saros.ui.preferencePages;
+package saros.ui.preference_pages;
 
 import org.eclipse.jface.preference.IPreferenceStore;
 import org.eclipse.jface.preference.PreferencePage;
